Move proxy to the separate module
