+++ src/main.rs
@@ -9,6 +9,7 @@ mod interceptor;
 mod jet;
 mod jet_client;
 mod logger;
+mod proxy;
 mod rdp;
 mod routing_client;
 mod transport;
@@ -16,43 +17,40 @@ mod utils;
 mod websocket_client;
 
 use std::collections::HashMap;
+use std::error::Error;
 use std::io;
 use std::io::ErrorKind;
 use std::net::{SocketAddr, ToSocketAddrs};
-use std::path;
-use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::atomic::AtomicU64;
 use std::sync::{Arc, Mutex};
 use std::time::Duration;
 
 use futures::{future, future::ok, future::Either, Future, Stream};
+use hyper::service::{make_service_fn, service_fn};
+use lazy_static::lazy_static;
+use saphir::server::HttpService;
+use slog::{o, Logger};
+use slog_scope::{error, info, slog_error};
+use slog_scope_futures::future01::FutureExt;
 use tokio::runtime::Runtime;
 use tokio::runtime::TaskExecutor;
+use tokio_rustls::{TlsAcceptor, TlsStream};
 use tokio_tcp::{TcpListener, TcpStream};
-
-use lazy_static::lazy_static;
-use slog::{o, Logger};
-use slog_scope::{error, info, slog_error, warn};
-use slog_scope_futures::future01::FutureExt;
 use url::Url;
+use x509_parser::pem::pem_to_der;
 
-use crate::config::{Config, Protocol};
+use crate::config::Config;
 use crate::http::http_server::HttpServer;
-use crate::interceptor::pcap::PcapInterceptor;
-use crate::interceptor::{rdp::RdpMessageReader, UnknownMessageReader, WaykMessageReader};
 use crate::jet_client::{JetAssociationsMap, JetClient};
+use crate::proxy::Proxy;
 use crate::rdp::RdpClient;
 use crate::routing_client::Client;
 use crate::transport::tcp::TcpTransport;
-use crate::transport::{JetTransport, Transport};
-use hyper::service::{make_service_fn, service_fn};
+use crate::transport::ws::{TcpWebSocketServerHandshake, TlsWebSocketServerHandshake, WsTransport};
+use crate::transport::JetTransport;
 use crate::utils::{get_pub_key_from_der, load_certs, load_private_key};
 use crate::websocket_client::{WebsocketService, WsClient};
-use std::error::Error;
-use crate::transport::ws::{WsTransport, TlsWebSocketServerHandshake, TcpWebSocketServerHandshake};
-use tokio_rustls::{TlsAcceptor, TlsStream};
-use saphir::server::HttpService;
 
-use x509_parser::pem::pem_to_der;
 
 const SOCKET_SEND_BUFFER_SIZE: usize = 0x7FFFF;
 const SOCKET_RECV_BUFFER_SIZE: usize = 0x7FFFF;
@@ -137,110 +135,6 @@ fn set_socket_option(stream: &TcpStream, logger: &Logger) {
 
     if let Err(e) = stream.set_recv_buffer_size(SOCKET_RECV_BUFFER_SIZE) {
         slog_error!(logger, "set_recv_buffer_size on TcpStream failed: {}", e);
-    }
-}
-
-struct Proxy {
-    config: Config,
-}
-
-impl Proxy {
-    pub fn new(config: Config) -> Self {
-        Proxy { config }
-    }
-
-    pub fn build<T: Transport, U: Transport>(
-        &self,
-        server_transport: T,
-        client_transport: U,
-    ) -> Box<dyn Future<Item=(), Error=io::Error> + Send> {
-        let jet_sink_server = server_transport.message_sink();
-        let mut jet_stream_server = server_transport.message_stream();
-
-        let jet_sink_client = client_transport.message_sink();
-        let mut jet_stream_client = client_transport.message_stream();
-
-        if let Some(pcap_files_path) = self.config.pcap_files_path() {
-            let server_peer_addr = jet_stream_server.peer_addr().unwrap();
-            let client_peer_addr = jet_stream_client.peer_addr().unwrap();
-
-            let filename = format!(
-                "{}({})-to-{}({})-at-{}.pcap",
-                client_peer_addr.ip(),
-                client_peer_addr.port(),
-                server_peer_addr.ip().to_string(),
-                server_peer_addr.port(),
-                chrono::Utc::now().format("%Y-%m-%d_%H-%M-%S")
-            );
-            let mut path = path::PathBuf::from(pcap_files_path);
-            path.push(filename);
-
-            let mut interceptor = PcapInterceptor::new(
-                server_peer_addr,
-                client_peer_addr,
-                path.to_str().expect("path to pcap files must be valid"),
-            );
-
-            match self.config.protocol() {
-                Protocol::WAYK => {
-                    info!("WaykMessageReader will be used to interpret application protocol.");
-                    interceptor.set_message_reader(WaykMessageReader::get_messages);
-                }
-                Protocol::RDP => {
-                    info!("RdpMessageReader will be used to interpret application protocol");
-                    interceptor.set_message_reader(RdpMessageReader::get_messages);
-                }
-                Protocol::UNKNOWN => {
-                    warn!("Protocol is unknown. Data received will not be split to get application message.");
-                    interceptor.set_message_reader(UnknownMessageReader::get_messages);
-                }
-            }
-
-            jet_stream_server.set_packet_interceptor(Box::new(interceptor.clone()));
-            jet_stream_client.set_packet_interceptor(Box::new(interceptor.clone()));
-        }
-
-        // Build future to forward all bytes
-        let f1 = jet_stream_server.forward(jet_sink_client);
-        let f2 = jet_stream_client.forward(jet_sink_server);
-
-        SESSION_IN_PROGRESS_COUNT.fetch_add(1, Ordering::Relaxed);
-
-        Box::new(f1.and_then(|(mut jet_stream, mut jet_sink)| {
-            // Shutdown stream and the sink so the f2 will finish as well (and the join future will finish)
-            let _ = jet_stream.shutdown();
-            let _ = jet_sink.shutdown();
-            ok((jet_stream, jet_sink))
-        })
-            .join(f2.and_then(|(mut jet_stream, mut jet_sink)| {
-                // Shutdown stream and the sink so the f2 will finish as well (and the join future will finish)
-                let _ = jet_stream.shutdown();
-                let _ = jet_sink.shutdown();
-                ok((jet_stream, jet_sink))
-            }))
-            .and_then(|((jet_stream_1, jet_sink_1), (jet_stream_2, jet_sink_2))| {
-                let server_addr = jet_stream_1
-                    .peer_addr()
-                    .map(|addr| addr.to_string())
-                    .unwrap_or_else(|| "unknown".to_string());
-                let client_addr = jet_stream_2
-                    .peer_addr()
-                    .map(|addr| addr.to_string())
-                    .unwrap_or_else(|| "unknown".to_string());
-                info!(
-                    "Proxy result : {} bytes read on {server} and {} bytes written on {client}. {} bytes read on {client} and {} bytes written on {server}",
-                    jet_stream_1.nb_bytes_read(),
-                    jet_sink_1.nb_bytes_written(),
-                    jet_stream_2.nb_bytes_read(),
-                    jet_sink_2.nb_bytes_written(),
-                    server = &server_addr,
-                    client = &client_addr
-                );
-                ok(())
-            }).then(|result| {
-            SESSION_IN_PROGRESS_COUNT.fetch_sub(1, Ordering::Relaxed);
-            result
-        }))
     }
 }
